Revert "m1721: wifi: Explicitly disable protected management frames"
This reverts the previous commit that explicitly disabled
Protected Management Frames (PMF) in the overlay.

The original fix was implemented to address compatibility issues
with WPA2/3 mixed access points, where the device would incorrectly
interact with PMF queries. However, after further review, it was found
that the `pmf=1` setting in the overlay was misunderstood, as it sets
PMF to be optional rather than explicitly enabling it.
This caused unnecessary disabling of PMF functionality for the device.

Reverting this change restores the correct behavior,
aligning with the intended configuration for WPA3 support,
without causing disruptions with PMF support.

This reverts commit 2dad75f00d5935e07e2414a17f4071ec8a0c2cdf.